[ci] Add B200 Kimi K3 integration lane
