+++ .github/pytorch-probot.yml
@@ -1,4 +1,5 @@
 ciflow_push_tags:
+  - ciflow/b200
   - ciflow/h100.8
   - ciflow/rl
 labeler_config: labeler.yml
