+++ .github/workflows/integration_test_8gpu_rl.yaml
@@ -3,13 +3,12 @@ name: RL Integration Tests
 on:
   push:
     branches: [ main ]
+    tags:
+      - ciflow/rl/*
     paths:
       - 'torchtitan/experiments/rl/**'
       - 'torchtitan/models/qwen3_5/**'
       - '.github/workflows/integration_test_8gpu_rl.yaml'
-  pull_request:
-    types: [labeled, synchronize]
-    branches: [ main ]
   schedule:
     # Runs every 12 hours
     - cron: '0 */12 * * *'
@@ -30,9 +29,7 @@ jobs:
   # Step 1: Dynamically compute the matrix based on conditions
   set-matrix:
     # Skip scheduled runs on forks, where they would only fail and email the fork owner
-    if: >-
-      (github.repository_owner == 'pytorch' || github.event_name != 'schedule') &&
-      (github.event_name != 'pull_request' || contains(github.event.pull_request.labels.*.name, 'ciflow/rl'))
+    if: github.repository_owner == 'pytorch' || github.event_name != 'schedule'
     uses: ./.github/workflows/set-matrix.yaml
     with:
       is-experimental: true
