+++ .github/workflows/integration_test_b200.yaml
@@ -1,0 +1,97 @@
+name: B200 Integration
+
+on:
+  push:
+    branches: [ main ]
+    tags:
+      - ciflow/b200/*
+    paths:
+      - 'torchtitan/models/kimi_k3/**'
+      - 'torchtitan_recipes/tests/b200.py'
+      - 'tests/integration_tests/b200.py'
+      - '.github/workflows/integration_test_b200.yaml'
+  pull_request:
+    paths:
+      - '.github/workflows/integration_test_b200.yaml'
+  workflow_dispatch:
+
+concurrency:
+  group: unit-test-${{ github.workflow }}-${{ github.ref == 'refs/heads/main' && github.run_number || github.ref }}
+  cancel-in-progress: true
+
+defaults:
+  run:
+    shell: bash -l -eo pipefail {0}
+
+permissions:
+  id-token: write
+  contents: read
+
+jobs:
+  build-test:
+    # Docker Hub credentials are unavailable to pull_request runs from forks.
+    # Those PRs run through the ciflow/b200 tag after workflow approval instead.
+    if: >-
+      github.repository_owner == 'pytorch' &&
+      (github.event_name != 'pull_request' || !github.event.pull_request.head.repo.fork)
+    runs-on: linux.dgx.b200.8
+    timeout-minutes: 60
+    container:
+      image: nvidia/cuda:12.9.1-devel-ubuntu24.04
+      options: --gpus all
+      credentials:
+        username: pytorchbot
+        password: ${{ secrets.DOCKER_HUB_READONLY_TOKEN }}
+    steps:
+      - name: Check out repo
+        uses: actions/checkout@v7
+
+      - name: Set up uv
+        uses: astral-sh/setup-uv@v6
+        with:
+          python-version: '3.12'
+          enable-cache: true
+
+      - name: Run B200 integration tests
+        run: |
+          set -eux
+
+          uv venv --python 3.12
+          source .venv/bin/activate
+
+          DRIVER_VERSION=$(nvidia-smi --query-gpu=driver_version --format=csv,noheader | head -n 1 || true)
+          echo "CUDA driver version: ${DRIVER_VERSION}"
+
+          uv pip install --pre torch torchvision \
+            --index-url https://download.pytorch.org/whl/nightly/cu129
+          uv pip install -r requirements.txt
+          uv pip install -r torchtitan/models/kimi_k3/requirements.txt
+
+          python - <<'PY'
+          import torch
+
+          capability = torch.cuda.get_device_capability()
+          if capability not in ((10, 0), (10, 3)):
+              raise RuntimeError(
+                  f"B200 integration requires SM100 or SM103, got SM{capability[0]}{capability[1]}"
+              )
+          PY
+
+          mkdir -p "$RUNNER_TEMP/artifacts-to-be-uploaded"
+
+          CUDA_HOME=/usr/local/cuda TORCH_SHOW_CPP_STACKTRACES=1 \
+            python -m tests.integration_tests.run_tests \
+            "$RUNNER_TEMP/artifacts-to-be-uploaded/b200" \
+            --test_suite b200 \
+            --execution_mode real_pg \
+            --gpu_arch_type cuda \
+            --ngpu 8
+
+          rm -rf "$RUNNER_TEMP/artifacts-to-be-uploaded/b200"/*/checkpoint
+
+      - name: Upload artifacts
+        if: always()
+        uses: actions/upload-artifact@v7
+        with:
+          name: b200-outputs
+          path: ${{ runner.temp }}/artifacts-to-be-uploaded
