+++ tests/integration_tests/b200.py
@@ -1,0 +1,21 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+import torchtitan_recipes.tests.b200 as recipes
+
+from tests.integration_tests import OverrideDefinitions
+
+
+def build_b200_tests_list() -> list[OverrideDefinitions]:
+    """Build integration tests that require B200-class hardware."""
+    return [
+        OverrideDefinitions(
+            configs=[recipes.kimi_k3_debugmodel_mm_fsdp2],
+            test_descr="Kimi K3 multimodal FSDP",
+            test_name="kimi_k3_mm_fsdp",
+            ngpu=2,
+        ),
+    ]
